backend/tree_data/requirements.txt
certifi==2018.11.29
chardet==3.0.4
DendroPy==4.4.0
Django==2.1.3
django-filter==2.0.0
django-treebeard==4.3
djangorestframework==3.9.0
idna==2.7
Markdown==3.0.1
python-nexus==1.63
pytz==2018.7
requests==2.20.1
urllib3==1.24.1
